move Item and suggest_name into lib mod
