+++ src/lib.rs
@@ -1,4 +1,6 @@
 #![doc = include_str!("../README.md")]
+
+use std::collections::BTreeMap;
 pub mod asm;
 pub mod cached_lines;
 pub mod demangle;
@@ -11,4 +13,51 @@ macro_rules! color {
     ($item:expr, $color:expr) => {
         owo_colors::OwoColorize::if_supports_color(&$item, owo_colors::Stream::Stdout, $color)
     };
+}
+
+#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq)]
+pub struct Item {
+    /// demangled name
+    pub name: String,
+    /// demangled name with hash
+    pub hashed: String,
+    /// sequential number of demangled name
+    pub index: usize,
+    /// number of lines
+    pub len: usize,
+}
+
+pub fn suggest_name(search: &str, full: bool, items: &[Item]) -> anyhow::Result<()> {
+    let names = items.iter().fold(BTreeMap::new(), |mut m, item| {
+        m.entry(if full { &item.hashed } else { &item.name })
+            .or_insert_with(Vec::new)
+            .push(item.len);
+        m
+    });
+
+    if names.is_empty() {
+        #[allow(clippy::redundant_else)]
+        if search.is_empty() {
+            anyhow::bail!("This target defines no functions")
+        } else {
+            anyhow::bail!("No matching functions, try relaxing your search request")
+        }
+    }
+
+    #[allow(clippy::cast_sign_loss)]
+    #[allow(clippy::cast_precision_loss)]
+    let width = (items.len() as f64).log10().ceil() as usize;
+
+    println!("Try one of those by name or a sequence number");
+    let mut ix = 0;
+    for (name, lens) in names.iter() {
+        println!(
+            "{ix:width$} {:?} {:?}",
+            color!(name, owo_colors::OwoColorize::green),
+            color!(lens, owo_colors::OwoColorize::cyan),
+        );
+        ix += lens.len();
+    }
+
+    std::process::exit(1);
 }
